+++ lib/character_map.rb
@@ -1,7 +1,7 @@
 
-class Charactermap
+module Charactermap
 
-  def character_map
+  def char_map
     ["a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o", "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z", "0", "1", "2", "3", "4", "5", "6", "7", "8", "9", " ", ",", ".", "!"]
   end
