+++ lib/cypher_module.rb
@@ -1,4 +1,7 @@
+require './lib/character_map'
+
 module Cyphers
+  include Charactermap
 
   def rotation_a
    @new_key.key_a + @new_offset.off_a
@@ -14,5 +17,12 @@ module Cyphers
 
   def rotation_d
      @new_key.key_d + @new_offset.off_d
+  end
+
+  def rotate_chars(char, index)
+    return char_map.rotate(rotation_a)[char_map.index(char)] if index == 0
+    return char_map.rotate(rotation_b)[char_map.index(char)] if index == 1
+    return char_map.rotate(rotation_c)[char_map.index(char)] if index == 2
+    return char_map.rotate(rotation_d)[char_map.index(char)] if index == 3
   end
 end
