--- secret_message.txt
@@ -1,1 +1,0 @@
-this is a secret message. numbers are 710
